miltest: Allow control of software polling to help hardware debug

+++ driver/mil1553P.h
@@ -161,7 +161,7 @@ struct working_area_s {
 	struct tx_queue_s tx_queue[MAX_DEVS];          /** Data and commands waiting to be transmitted */
 	uint32_t icnt;                                 /** Total interrupt count */
 	uint32_t isrdebug;                             /** Trace ISR */
-	uint32_t nopol;                                /** No polling flag */
+	unsigned long nopol;                           /** No polling flag */
 };
 
 #endif
